bookmarklet with Javascript.
A bookmarklet is a bookmark stored in a web browser that contains
JavaScript code to extend the browser’s functionality. When you click on
the bookmark in the bookmarks or favorites bar of your browser, the
JavaScript code is executed on the website being displayed in the browser.
This is very useful for building tools that interact with other websites.
Some online services, such as Pinterest, implement their own bookmarklet
to let users share content from other sites on their platforms