Added asynchronous actions with JavaScript
Add a like button to the image detail page to let users click
on it to like an image. When users click the like button, we will send an
HTTP request to the web server using JavaScript. This will perform the like
action without reloading the whole page. For this functionality, we will
implement a view that allows users to like/unlike images