Creating an activity stream application
Many social websites display an activity stream to their users so that they
can track what other users do on the platform. An activity stream is a list of
recent activities performed by a user or a group of users. For example,
Facebook’s News Feed is an activity stream. Sample actions can be user X
bookmarked image Y or user X is now following user Y.
I build an activity stream application so that every user can
see the recent interactions of the users they follow. To do so, you will need a
model to save the actions performed by users on the website and a simple
way to add actions to the feed